requirements.txt
llama-index-0.10.56
llama-index-llms-together-0.1.3
instructor
matplotlib
openai
streamlit-1.36.0
python-dotenv-1.0.1
graphviz